Development

+++ ParkUp.API/Models/ParkingSpaceDTO.cs
@@ -14,7 +14,7 @@ namespace ParkUp.API.Models
         public string Name { get; set; }
         public string StreetName { get; set; }
         public string Description { get; set; }
-        public double HourlyPrice { get; set; }
+        public decimal HourlyPrice { get; set; }
         public bool IsActive { get; set; } = false;
         public bool IsTaken { get; set; } = false;
         public DateTime DateAdded { get; set; }
